\name{Sensitivity_list}
\alias{Sensitivity_list}
\docType{data}
\title{
Example Data Set for the Function carlit().
}
\description{
A reference list for sensitivity levels of macroalgal communities.
}
\usage{data(Sensitivity_list)}
\format{
  A data frame with 19 observations on the following 2 variables.
  \describe{
    \item{\code{Community}}{a factor with levels \code{Blue_greens} \code{Corallina} \code{Cymodocea_nodosa} \code{Cystoseira_balearica} \code{Cystoseira_compressa} \code{Cystoseira_crinita} \code{Cystoseira_mediterranea_1} \code{Cystoseira_mediterranea_2} \code{Cystoseira_mediterranea_3} \code{Cystoseira_mediterranea_4} \code{Cystoseira_mediterranea_5} \code{Cystoseira_sheltered} \code{Encrusting_corallines} \code{Green_algae} \code{Haliptilon} \code{Mytilus} \code{Posidonia_reef} \code{Trottoir} \code{Zostera_noltii}}
    \item{\code{SL}}{a numeric vector}
  }
}
\examples{
data(Sensitivity_list)
str(Sensitivity_list)
}
\keyword{datasets}
\keyword{CARLIT}
\keyword{carlit}